Split session
